+++ src/routes/index.tsx
@@ -1,6 +1,7 @@
 import { createBrowserRouter } from 'react-router-dom';
 
 import ReviewDetail from '@/pages/ReviewDetailPage';
+import SignInPage from '@/pages/SignInPage/SignInPage';
 
 import MainPage from '../pages/MainPage';
 import NotificationPage from '../pages/NotificationPage';
@@ -8,7 +9,6 @@ import ProfilePage from '../pages/ProfilePage';
 import ReviewList from '../pages/ReviewList';
 import ReviewPage from '../pages/ReviewPage';
 import SearchPage from '../pages/SearchPage';
-import SignInPage from '../pages/SignInPage';
 import SignUpPage from '../pages/SignUpPage';
 import TestPage from '../pages/TestPage';
 import { PATH } from './path';
